[AC-1325] Addressing feedback
