+++ src/selectors/accessConditionState.js
@@ -1,6 +1,5 @@
 import { createSelector } from 'reselect';
 import _ from 'lodash';
-import { formatToMatchAccountPlan } from 'src/helpers/billing';
 
 const getAccount = state => state.account;
 const getUser = state => state.currentUser;
@@ -8,15 +7,23 @@ const getBundles = state => _.get(state, 'billing.bundles', []);
 const getBundlePlans = state => _.get(state, 'billing.bundlePlans', []);
 const getACReady = state => state.accessControlReady;
 const getBillingSubscription = state => _.get(state, 'billing.subscription', {});
-
+const currentFreePlans = ['free500-1018', 'free15K-1018', 'free500-0419', 'free500-SPCEU-0419'];
 export const getCurrentAccountPlan = createSelector(
   [getAccount, getBundlePlans, getBundles, getBillingSubscription],
   (account, bundlePlans, bundles, subscription) => {
-    return formatToMatchAccountPlan({
+    const currentPlan = {
       ...bundlePlans.find(plan => plan.plan === account.subscription.code),
       ...bundles.find(bundle => bundle.bundle === account.subscription.code),
       products: subscription.products,
-    });
+    };
+    return {
+      billingId: currentPlan.billing_id,
+      code: currentPlan.plan,
+      includesIp: !currentPlan.status ? false : true,
+      isFree: currentFreePlans.includes(currentPlan.plan),
+      status: !currentPlan.status ? 'deprecated' : currentPlan.status, //since bundlePlans don't return deprecated plans;
+      ...currentPlan,
+    };
   },
 );
